fix: prevent memory spike when switching from EPUB to large manga
- Synchronously set `isLoading` to true in Redux when changing books. This ensures the `NovelReader` component is unmounted immediately, preventing it from incorrectly attempting to load large non-EPUB files during navigation race conditions.
- Moved the `CircularProgress` (loading spinner) from `ComicReader` to the parent `BookReader` component for consistent loading states across different reader types.
- Added a defensive file extension check in the `useNovelReader` hook to skip processing for non-EPUB files.
- Improved resource cleanup in `useNovelReader` by properly tracking and destroying the `Book` instance.
