+++ docs/ja_JP/CHANGELOG.md
@@ -24,6 +24,8 @@
 * 起動時の初期ビューの設定で、誤ったアイコンが表示されていた問題を修正（Language アイコンから RocketLaunch アイコンへ変更）(#184)
 * コンポーネントの遅延読み込み（`React.lazy`）に伴う `<Suspense>` の欠落が原因でアプリがクラッシュする可能性を修正し、静的インポートへの切り替えによって初期表示時のUXを改善 (#185)
 * 前回開いた本を復元する初期化処理の際、`getRecentlyReadBooks` API が意図せず重複して呼び出されていた不具合を修正 (#187)
+* EPUB小説を読んだ直後に大容量の漫画ファイルを開くと、メモリ使用量が急増してアプリが強制終了する問題を修正 (#188)
+
 
 ## [2.0.2] - 2026-04-15
