+++ src/features/BookReader/hooks/useNovelReader.ts
@@ -55,6 +55,7 @@ export const useNovelReader = ({ filePath }: UseNovelReaderOptions) => {
   const theme = useAppTheme();
   const viewerRef = useRef<HTMLDivElement>(null);
   const viewRef = useRef<View | null>(null);
+  const bookRef = useRef<Book | null>(null);
   const index = useAppSelector((state) => state.read.containerFile.index);
   const cfi = useAppSelector((state) => state.read.containerFile.cfi);
   const readingDirection = useAppSelector((state) => state.settings.reader.comic.readingDirection);
@@ -150,6 +151,14 @@ export const useNovelReader = ({ filePath }: UseNovelReaderOptions) => {
       viewRef.current?.close();
       viewRef.current?.remove();
       viewRef.current = null;
+      bookRef.current?.destroy?.();
+      bookRef.current = null;
+
+      // Defensive check: only read if the file is an EPUB.
+      if (!filePath.toLowerCase().endsWith(".epub")) {
+        error(`Attempted to load non-EPUB file in NovelReader: ${filePath}`);
+        return;
+      }
 
       const binaryData = await readFile(filePath);
 
@@ -159,6 +168,7 @@ export const useNovelReader = ({ filePath }: UseNovelReaderOptions) => {
 
       const file = new File([binaryData], filePath, { type: "application/epub+zip" });
       const book = await makeBook(file);
+      bookRef.current = book;
 
       if (!isMounted) {
         book.destroy?.();
@@ -246,6 +256,8 @@ export const useNovelReader = ({ filePath }: UseNovelReaderOptions) => {
       viewRef.current?.close();
       viewRef.current?.remove();
       viewRef.current = null;
+      bookRef.current?.destroy?.();
+      bookRef.current = null;
     };
   }, [filePath]);
